backend/requirements/base.txt
django>=5.0,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.3,<6.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
django-filter>=24.0,<25.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0